Worked on the JS, HTML and some CSS
Added spans to the lines that I needed Axios to change -have not solved
Made the card group smaller- still a large space after Sunday
Axios works for current date and weather search- idle state, wind, humidity and description are not being called. I feel like my showTemperature is causing that issue.
Line 60 is telling me that showCity is not a functions when it is, so there must be an issue there as well.